Move Merge/Transplant classes to the api.v1.params package
These classes are not used in user-facing java client methods.

They are used only as parameters to the REST API interfaces, so it
makes sense to keep them under the params package. This will also
improve isolation between API v1 and v2.

This change is expected to be backward-compatible to existing java
and REST API clients.
