+++ clients/client/src/main/java/org/projectnessie/client/http/v1api/HttpMergeReference.java
@@ -15,11 +15,11 @@
  */
 package org.projectnessie.client.http.v1api;
 
+import org.projectnessie.api.v1.params.ImmutableMerge;
 import org.projectnessie.client.builder.BaseMergeReferenceBuilder;
 import org.projectnessie.client.http.NessieApiClient;
 import org.projectnessie.error.NessieConflictException;
 import org.projectnessie.error.NessieNotFoundException;
-import org.projectnessie.model.ImmutableMerge;
 import org.projectnessie.model.MergeResponse;
 
 final class HttpMergeReference extends BaseMergeReferenceBuilder {
